+++ api/controllers/state/remote.js
@@ -44,7 +44,7 @@ module.exports = {
             // Block this request if we are changing states right now
             if (Meta['A'].changingState !== null)
                 return exits.error(new Error(`The system is in the process of changing states. The request was blocked to prevent clashes.`));
-            
+
             // Lock so that other state changing requests get blocked until we are done.
             await Meta.changeMeta({changingState: `Switching to remote`});
 
@@ -77,7 +77,7 @@ module.exports = {
                 await sails.helpers.songs.queue(sails.config.custom.subcats.IDs, 'Bottom', 1);
                 Status.errorCheck.prevID = moment();
                 await sails.helpers.error.count('stationID');
-                
+
                 // Queue a show opener if there is one
                 if (typeof sails.config.custom.showcats[Meta['A'].show] !== 'undefined')
                 {
@@ -85,9 +85,22 @@ module.exports = {
                 } else {
                     await sails.helpers.songs.queue([sails.config.custom.showcats["Default"]["Show Openers"]], 'Bottom', 1);
                 }
-                
+
                 await sails.helpers.rest.cmd('EnableAssisted', 0);
-                await Meta.changeMeta({queueFinish: moment().add(await sails.helpers.songs.calculateQueueLength(), 'seconds').toISOString(true), state: 'automation_remote', show: inputs.showname, topic: inputs.topic, trackStamp: null, lastID: moment().toISOString(true), webchat: inputs.webchat, djcontrols: this.req.payload.host});
+
+                var queueLength = await sails.helpers.songs.calculateQueueLength();
+
+                if (queueLength >= sails.config.custom.queueCorrection.remote)
+                {
+                    await sails.helpers.rest.cmd('EnableAutoDJ', 0); // Try to Disable autoDJ again in case it was mistakenly still active
+                    //await sails.helpers.songs.remove(true, sails.config.custom.subcats.noClearShow, false, false);
+                    if ((sails.config.custom.subcats.noClearShow && sails.config.custom.subcats.noClearShow.indexOf(Meta['A'].trackIDSubcat) === -1))
+                        await sails.helpers.rest.cmd('PlayPlaylistTrack', 1); // Skip currently playing track if it is not a noClearShow track
+
+                    queueLength = await sails.helpers.songs.calculateQueueLength();
+                }
+
+                await Meta.changeMeta({queueFinish: moment().add(queueLength, 'seconds').toISOString(true), state: 'automation_remote', show: inputs.showname, topic: inputs.topic, trackStamp: null, lastID: moment().toISOString(true), webchat: inputs.webchat, djcontrols: this.req.payload.host});
             } else {
                 // Otherwise, just update metadata but do not do anything else
                 await Meta.changeMeta({show: inputs.showname, topic: inputs.topic, trackStamp: null, webchat: inputs.webchat, djcontrols: this.req.payload.host});